+++ Sources/OnboardingKit/_Deprecated/Views+Deprecated.swift
@@ -1,0 +1,22 @@
+import SwiftUI
+
+public extension OnboardingSlideshow {
+
+    @available(*, deprecated, message: "Apply a background to the view instead.")
+    init<Background: View> (
+        pages: [Page],
+        pageIndex: Binding<Int>,
+        config: OnboardingSlideshowConfiguration = .standard,
+        onStoryCompleted: @escaping () -> Void,
+        @ViewBuilder background: @escaping (PageInfo) -> Background,
+        @ViewBuilder content: @escaping ContentBuilder
+    ) {
+        self.init(
+            pages: pages,
+            pageIndex: pageIndex,
+            config: config,
+            onStoryCompleted: onStoryCompleted,
+            content: content
+        )
+    }
+}